fix medusa model relationships
