+++ apps/backend/src/modules/quote/models/quote.ts
@@ -1,5 +1,4 @@
 import { model } from "@medusajs/framework/utils";
-import { Message } from "./message";
 
 export const Quote = model.define("quote", {
   id: model.id({ prefix: "quo" }).primaryKey(),
@@ -17,4 +16,13 @@ export const Quote = model.define("quote", {
   order_change_id: model.text(),
   cart_id: model.text(),
   messages: model.hasMany(() => Message),
+});
+
+export const Message = model.define("message", {
+  id: model.id({ prefix: "mess" }).primaryKey(),
+  text: model.text(),
+  item_id: model.text().nullable(),
+  admin_id: model.text().nullable(),
+  customer_id: model.text().nullable(),
+  quote: model.belongsTo(() => Quote, { mappedBy: "messages" }),
 });
